+++ ChangeLog.md
@@ -1,3 +1,6 @@
+## Changes between ifm3d-ros 0.4.2 and 0.5.0
+* Converted primary data publisher to a nodelet architecture
+
 ## Changes between ifm3d-ros 0.4.1 and 0.4.2
 * Now requires ifm3d 0.9.0 and by association the more modernized tooling
   (C++14, cmake 3.5, dropped support for 14.04/Indigo, etc.)